feat: massive overhaul, ESM-only support
I have attempted this 5-6 times over the last year, across 10+ different
branches, and it is finally building.

Tests are certainly not passing, but the tool now bootstraps itself and compiles
to split entry points with full ESM compatibility, and no CJS support.

This is the theoretical lower bound for module size: Everything except each
entry-point is external, i.e. relative and module imports are preserved. As an
input to client bundles, it is maximally "tree-shakeable" for AOT static
compilation but also at runtime (each import path will only execute as needed
ipso facto).